Added support for 64-bit build and OpenSSL 1.1. and fixed Linux build failures

+++ TSS.CPP/Src/Extensions/TPMS_PCR_SELECTION.h
@@ -72,7 +72,7 @@ public:
 std::vector<UINT32> ToArray()
 {
     std::vector<UINT32> arr;
-    int maxIs = pcrSelect.size() * 8;
+    int maxIs = (int)pcrSelect.size() * 8;
 
     for (int j = 0; j < maxIs; j++) {
         if (PcrIsSelected(j)) {
